Removes carousel versions
