+++ src/Apps/Artwork/Components/ArtworkImageBrowser/ArtworkImageBrowser.tsx
@@ -1,6 +1,6 @@
 import { ArtworkImageBrowser_artwork } from "__generated__/ArtworkImageBrowser_artwork.graphql"
 import { Lightbox } from "Components/v2"
-import { BaseCarousel as Carousel } from "Components/v2/CarouselV3"
+import { BaseCarousel as Carousel } from "Components/v2/Carousel"
 import React from "react"
 import styled from "styled-components"
 import { Media } from "Utils/Responsive"
@@ -10,7 +10,7 @@ import { Box, ChevronIcon, Col, color, Flex, space } from "@artsy/palette"
 interface ArtworkBrowserProps {
   imageAlt: string
   images: ArtworkImageBrowser_artwork["images"]
-  setFlickityRef: (flickityRef: Flickity) => void
+  setCarouselRef: (carouselRef: Carousel) => void
 }
 
 export const ArtworkImageBrowser = (props: ArtworkBrowserProps) => {
@@ -31,7 +31,7 @@ export class LargeArtworkImageBrowser extends React.Component<
 > {
   render() {
     const hasMultipleImages = this.props.images.length > 1
-    const { imageAlt, images, setFlickityRef } = this.props
+    const { imageAlt, images, setCarouselRef } = this.props
 
     // FIXME: During SSR pass want to hide other images. Work around for lack
     // of SSR support in Flickity.
@@ -53,7 +53,7 @@ export class LargeArtworkImageBrowser extends React.Component<
           options={options}
           oneSlideVisible
           height="60vh"
-          setFlickityRef={setFlickityRef}
+          setCarouselRef={setCarouselRef}
           data={carouselImages}
           renderLeftArrow={({ flickity }) => (
             <Col sm={1}>
@@ -101,7 +101,7 @@ export class SmallArtworkImageBrowser extends React.Component<
   ArtworkBrowserProps
 > {
   render() {
-    const { images, imageAlt, setFlickityRef } = this.props
+    const { images, imageAlt, setCarouselRef } = this.props
     // FIXME: During SSR pass want to hide other images. Work around for lack
     // of SSR support in Flickity.
     const carouselImages = typeof window === "undefined" ? [images[0]] : images
@@ -119,7 +119,7 @@ export class SmallArtworkImageBrowser extends React.Component<
           options={options}
           data={carouselImages}
           oneSlideVisible
-          setFlickityRef={setFlickityRef}
+          setCarouselRef={setCarouselRef}
           render={image => {
             return (
               <Flex
