+++ src/Components/Artist/ArtistCollectionsRail/ArtistCollectionsRail.tsx
@@ -2,7 +2,7 @@ import { Box, Sans, Spacer } from "@artsy/palette"
 import { ArtistCollectionsRail_collections } from "__generated__/ArtistCollectionsRail_collections.graphql"
 import { track } from "Artsy/Analytics"
 import * as Schema from "Artsy/Analytics/Schema"
-import { ArrowButton, Carousel } from "Components/v2/CarouselV3"
+import { ArrowButton, Carousel } from "Components/v2/Carousel"
 import { once } from "lodash"
 import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
@@ -53,7 +53,7 @@ export class ArtistCollectionsRail extends React.Component<
           <Spacer pb={1} />
 
           <Carousel
-            height={200}
+            height="200px"
             options={{
               groupCells: 1,
               wrapAround: true,
