+++ src/Components/v2/RecentlyViewed.tsx
@@ -6,7 +6,7 @@ import { track } from "Artsy/Analytics"
 import * as Schema from "Artsy/Analytics/Schema"
 import { renderWithLoadProgress } from "Artsy/Relay/renderWithLoadProgress"
 import { FillwidthItem } from "Components/Artwork/FillwidthItem"
-import { Carousel } from "Components/v2/CarouselV3"
+import { Carousel } from "Components/v2/Carousel"
 import React, { useContext } from "react"
 import { QueryRenderer } from "react-relay"
 import { createFragmentContainer, graphql } from "react-relay"
