+++ src/Components/v2/__stories__/Carousel.story.tsx
@@ -1,163 +1,207 @@
-import { Box, Image } from "@artsy/palette"
+import { Box, Flex, Image } from "@artsy/palette"
 import { artworkBricks, images } from "Apps/__tests__/Fixtures/Carousel"
 import { FillwidthItem } from "Components/Artwork/FillwidthItem"
-import { Carousel, LargeCarousel, SmallCarousel } from "Components/v2/Carousel"
 import { RelayStubProvider } from "DevTools/RelayStubProvider"
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
-import { Section } from "Utils/Section"
 
-storiesOf("Styleguide/Components", module).add("Carousel", () => {
+import { Carousel, LargeCarousel, SmallCarousel } from "Components/v2/Carousel"
+
+storiesOf("Styleguide/Components/Carousel", module)
+  .add("Responsive Carousel", () => {
+    return (
+      <Container>
+        <Carousel
+          data={images}
+          render={props => {
+            return (
+              <Image
+                px={5}
+                src={props.resized.url}
+                width={props.resized.width}
+                height={props.resized.height}
+              />
+            )
+          }}
+        />
+      </Container>
+    )
+  })
+
+  .add("Small Carousel", () => {
+    return (
+      <Container width="50%">
+        <SmallCarousel
+          data={images}
+          render={props => {
+            return (
+              <Image
+                px={5}
+                src={props.resized.url}
+                width={props.resized.width}
+                height={props.resized.height}
+              />
+            )
+          }}
+        />
+      </Container>
+    )
+  })
+
+  .add("Large Carousel", () => {
+    return (
+      <Container>
+        <LargeCarousel
+          data={images}
+          height="300px"
+          render={props => {
+            return (
+              <Image
+                px={5}
+                src={props.resized.url}
+                width={props.resized.width}
+                height={props.resized.height}
+              />
+            )
+          }}
+        />
+      </Container>
+    )
+  })
+
+  .add("Large Carousel with WrapAround", () => {
+    return (
+      <Container width="70%">
+        <LargeCarousel
+          showArrows
+          data={images}
+          height="300px"
+          options={{ wrapAround: true }}
+          render={props => {
+            return (
+              <Image
+                px={5}
+                src={props.resized.url}
+                width={props.resized.width}
+                height={props.resized.height}
+              />
+            )
+          }}
+        />
+      </Container>
+    )
+  })
+
+  .add("Carousel with ArtworkBricks", () => {
+    return (
+      <Container>
+        <RelayStubProvider>
+          <Carousel
+            data={artworkBricks}
+            render={artwork => {
+              const {
+                node: {
+                  image: { aspect_ratio },
+                },
+              } = artwork
+              return (
+                <FillwidthItem
+                  artwork={artwork.node}
+                  targetHeight={200}
+                  imageHeight={200}
+                  width={200 * aspect_ratio}
+                  margin={20}
+                />
+              )
+            }}
+          />
+        </RelayStubProvider>
+      </Container>
+    )
+  })
+
+  .add("Custom arrows with defaults", () => {
+    return (
+      <Container>
+        <Carousel
+          data={images}
+          render={props => {
+            return (
+              <Image
+                px={5}
+                src={props.resized.url}
+                width={props.resized.width}
+                height={props.resized.height}
+              />
+            )
+          }}
+          renderLeftArrow={({ Arrow }) => {
+            return (
+              <Box top={10} position="relative" bg="black10">
+                <Arrow />
+              </Box>
+            )
+          }}
+          renderRightArrow={({ Arrow }) => {
+            return (
+              <Box top={10} position="relative" bg="black10">
+                <Arrow />
+              </Box>
+            )
+          }}
+        />
+      </Container>
+    )
+  })
+
+  .add("Custom arrows", () => {
+    return (
+      <Container>
+        <Carousel
+          data={images}
+          render={props => {
+            return (
+              <Image
+                px={5}
+                src={props.resized.url}
+                width={props.resized.width}
+                height={props.resized.height}
+              />
+            )
+          }}
+          renderLeftArrow={({ flickity }) => {
+            return (
+              <Box
+                onClick={() => {
+                  flickity.previous()
+                }}
+              >
+                Prev
+              </Box>
+            )
+          }}
+          renderRightArrow={({ flickity }) => {
+            return (
+              <Box
+                onClick={() => {
+                  flickity.next()
+                }}
+              >
+                Next
+              </Box>
+            )
+          }}
+        />
+      </Container>
+    )
+  })
+
+const Container = ({ children, ...props }) => {
   return (
-    <>
-      <Section title="Responsive Carousel">
-        <Box width="70%">
-          <Carousel
-            data={images}
-            render={props => {
-              return (
-                <Image
-                  px={5}
-                  src={props.resized.url}
-                  width={props.resized.width}
-                  height={props.resized.height}
-                />
-              )
-            }}
-          />
-        </Box>
-      </Section>
-      <Section title="Small Carousel">
-        <Box width="50%">
-          <SmallCarousel
-            data={images}
-            render={props => {
-              return (
-                <Image
-                  px={5}
-                  src={props.resized.url}
-                  width={props.resized.width}
-                  height={props.resized.height}
-                />
-              )
-            }}
-          />
-        </Box>
-      </Section>
-      <Section title="Large Carousel">
-        <Box width="70%">
-          <LargeCarousel
-            data={images}
-            height={300}
-            render={props => {
-              return (
-                <Image
-                  px={5}
-                  src={props.resized.url}
-                  width={props.resized.width}
-                  height={props.resized.height}
-                />
-              )
-            }}
-          />
-        </Box>
-      </Section>
-      <Section title="Artwork Brick">
-        <Box width="70%">
-          <RelayStubProvider>
-            <Carousel
-              data={artworkBricks}
-              render={artwork => {
-                const {
-                  node: {
-                    image: { aspect_ratio },
-                  },
-                } = artwork
-
-                return (
-                  <FillwidthItem
-                    artwork={artwork.node}
-                    targetHeight={200}
-                    imageHeight={200}
-                    width={200 * aspect_ratio}
-                    margin={20}
-                  />
-                )
-              }}
-            />
-          </RelayStubProvider>
-        </Box>
-      </Section>
-      <Section title="Custom Arrows with defaults">
-        <Box width="70%">
-          <Carousel
-            data={images}
-            render={props => {
-              return (
-                <Image
-                  px={5}
-                  src={props.resized.url}
-                  width={props.resized.width}
-                  height={props.resized.height}
-                />
-              )
-            }}
-            renderLeftArrow={({ Arrow }) => {
-              return (
-                <Box top={10} position="relative" bg="black10">
-                  <Arrow />
-                </Box>
-              )
-            }}
-            renderRightArrow={({ Arrow }) => {
-              return (
-                <Box top={10} position="relative" bg="black10">
-                  <Arrow />
-                </Box>
-              )
-            }}
-          />
-        </Box>
-      </Section>
-      <Section title="Custom Arrows ">
-        <Box width="70%">
-          <Carousel
-            data={images}
-            render={props => {
-              return (
-                <Image
-                  px={5}
-                  src={props.resized.url}
-                  width={props.resized.width}
-                  height={props.resized.height}
-                />
-              )
-            }}
-            renderLeftArrow={({ getSlick }) => {
-              return (
-                <Box
-                  onClick={() => {
-                    getSlick().slickPrev()
-                  }}
-                >
-                  Prev
-                </Box>
-              )
-            }}
-            renderRightArrow={({ getSlick }) => {
-              return (
-                <Box
-                  onClick={() => {
-                    getSlick().slickNext()
-                  }}
-                >
-                  Next
-                </Box>
-              )
-            }}
-          />
-        </Box>
-      </Section>
-    </>
+    <Flex width="100%" justifyContent="center">
+      <Box width="70%" {...props}>
+        {children}
+      </Box>
+    </Flex>
   )
-})
+}
